Replace GUI with new implemented GUI
